feat: add sqlite schema migrations
